backend/requirements.txt
fastapi
uvicorn
python-dotenv
litellm
python-multipart
pypdf
httpx
